:recycle: refactoring excel.helpers into a package
