+++ django-data/image/excel/helpers/exceptions.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Jul  5 16:39:30 2019
+
+@author: Paolo Cozzi <cozzi@ibba.cnr.it>
+"""
+
+
+# A class to deal with template import errors
+class ExcelImportError(Exception):
+    pass
